% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{apply_handle_splits}
\alias{apply_handle_splits}
\title{Apply handle_splits() to RSA-911 dataset.}
\usage{
apply_handle_splits(data, special_cols, sep = ";")
}
\arguments{
\item{data}{An RSA-911 dataset.}

\item{special_cols}{A character vector of the names of variables with special
character values in the dataset.}

\item{sep}{The special character that is separating values within the
variable.}
}
\value{
The cleaned RSA-911 dataset, with new variables that contain the
cleaned, separated values without special characters.
}
\description{
This function cleans variables with special characters by splitting values
into separate variables that are separated by special characters by
applying the function within a dataset.
}
