% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_provider.R
\name{clean_provider}
\alias{clean_provider}
\title{Clean TRT Scores Data, with Provider Variables Focus}
\usage{
clean_provider(
  data,
  state_filter = NULL,
  id_col = NULL,
  remove_questions = TRUE,
  condense = FALSE,
  clean_id = TRUE
)
}
\arguments{
\item{data}{The TRT scores dataset.}

\item{state_filter}{A character vector identifying the state(s) of interest.
Defaults to NULL.}

\item{id_col}{Differing variable naming structure for participant ID.
(Eg. "X", or another name not similar to "participant" or "ID".)
Defaults to NULL.}

\item{remove_questions}{Defaults to TRUE. When TRUE, removes question item
variables.}

\item{condense}{Defaults to FALSE. When true, take medians across
participants in order to keep only unique combinations of provider and
service.}

\item{clean_id}{TRUE or FALSE. Defaults to TRUE, when TRUE, rows where
participant ID is missing are removed.}
}
\value{
A cleaned data frame, including only provider-relevant variables,
restructured.
}
\description{
This function cleans a scores dataset, based on the standard data
structure, returning a dataset focused on provider-variable analysis.
}
