% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_utah.R
\name{clean_utah}
\alias{clean_utah}
\title{Clean Utah RSA-911 Data}
\usage{
clean_utah(
  data,
  aggregate = TRUE,
  unidentified_to_0 = TRUE,
  remove_desc = TRUE,
  remove_strictly_na = TRUE,
  clean_specials = NULL
)
}
\arguments{
\item{data}{The Utah RSA-911 dataset.}

\item{aggregate}{TRUE or FALSE. Defaults to TRUE, when TRUE, rows are
aggregated to include only unique combinations of participants, year,
and quarter.}

\item{unidentified_to_0}{TRUE or FALSE. Defaults to TRUE, when TRUE,
variables where unidentified is represented by 9,
values are converted to 0.}

\item{remove_desc}{TRUE or FALSE. Defaults to TRUE, when TRUE, description
variables are removed to minimize redundancy.}

\item{remove_strictly_na}{TRUE or FALSE. Defaults to TRUE, when TRUE,
variables that contain only NA values are removed.}

\item{clean_specials}{A character vector of name(s) of variables with special
characters to be cleaned. New, separate variables for each value space will
be appended to end of dataset with the the following naming convention:
original_variable_name_Place1, original_variable_name_Place2, etc.
Defaults to NULL. Note that if many are listed, cleaning process will be
very slow.}
}
\value{
A cleaned data frame, with incorrect, blank, messy values replaced,
additional, helpful variables created, and unnecessary variables removed.
}
\description{
This function cleans a Utah RSA-911 dataset, based on the standard data
structure.
}
