% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Create Metadata}
\usage{
create_metadata(data, includes_scores = TRUE)
}
\arguments{
\item{data}{The merged or RSA-911 dataset.}

\item{includes_scores}{TRUE or FALSE. Defaults to TRUE, when TRUE, metadata
creation includes process for scores data variables--inputted data frame is
treated as a merged dataset. When FALSE, metadata process only includes
steps relevant for an RSA-911 dataset.}
}
\value{
A data frame with one row per participant.
}
\description{
This function aggregates a merged or RSA-911 dataset to contain only one row
per unique participant, to make analysis and modeling more intuitive. The
aggregation process includes condensing rows by taking medians for numeric
variables, keeping most common values for character and factor variables,
and keeping most recent values for date-related variables.
}
