% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_abbrev}
\alias{handle_abbrev}
\title{CLean/abbreviate name variables.}
\usage{
handle_abbrev(x)
}
\arguments{
\item{x}{The name variable.}
}
\value{
The cleaned name variable, with abbreviated values.
}
\description{
This function cleans and shortens name variables, for example, to applied on
a variable such as Provider.
}
