% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_age_group}
\alias{handle_age_group}
\title{Create age group column in RSA-911 dataset.}
\usage{
handle_age_group(x)
}
\arguments{
\item{x}{An RSA-911 age at application variable to be grouped.}
}
\value{
The new age group values, see above list.
}
\description{
This function groups variable Age at Application into ordered levels:
<5
5-7
8-10
11-13
14-16
17-19
20-22
23-25
26-30
31-40
41+
}
