% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_blanks}
\alias{handle_blanks}
\title{Clean a variable with blank values.}
\usage{
handle_blanks(x)
}
\arguments{
\item{x}{An RSA-911 variable.}
}
\value{
The cleaned variable, where blanks and NULLs are replaced with 0s.
}
\description{
This function cleans a RSA-911 variable with blanks and NULL values.
}
