% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_code}
\alias{handle_code}
\title{Convert character coded variables.}
\usage{
handle_code(x)
}
\arguments{
\item{x}{A variable containing code values, so values that have little
meaning in numeric or factor form. Simply represent recorded ids.}
}
\value{
The cleaned code variable.
}
\description{
This function converts a RSA-911 date variable written in order YYYYMMDD as
one number to the appropriate date.
}
