% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_date}
\alias{handle_date}
\title{Clean date variable.}
\usage{
handle_date(x)
}
\arguments{
\item{x}{A date variable, written as a numeric YYYYMMDD.}
}
\value{
The converted date variable.
}
\description{
This function converts a RSA-911 date variable written in order YYYYMMDD as
one number to the appropriate date.
}
