% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_excel_date}
\alias{handle_excel_date}
\title{Clean Excel-based date variable.}
\usage{
handle_excel_date(x)
}
\arguments{
\item{x}{An Excel date variable, origin 1899-12-30.}
}
\value{
The converted date variable.
}
\description{
This function converts a RSA-911 Excel-based date variable written as one
number with origin date 1899-12-30 to the appropriate date.
}
