% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_impairment_group}
\alias{handle_impairment_group}
\title{Create Primary Impairment and Secondary Impairment grouping columns in
RSA-911 dataset.}
\usage{
handle_impairment_group(x)
}
\arguments{
\item{x}{An RSA-911 impairment (or cause) variable to be grouped.}
}
\value{
The new variable with grouped character values Visual, Aud_Comm,
Physical, Intell_Learn, and Psych, rather than values 0-19.
}
\description{
This function groups impairments (values 0-19) by the codebook definitions
(visual, auditory/communicative, physical, intellectual/learning,
psychological). This process
}
