% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_mixed_date}
\alias{handle_mixed_date}
\title{Clean mixed-format date variable.}
\usage{
handle_mixed_date(x)
}
\arguments{
\item{x}{An date variable (if Excel date, origin 1899-12-30).}
}
\value{
The converted date variable.
}
\description{
This function converts a RSA-911 date variable with different entered
formats.
}
