% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_nines}
\alias{handle_nines}
\title{Convert date variable.}
\usage{
handle_nines(x, unidentified_to_0 = TRUE)
}
\arguments{
\item{x}{A date variable, written as a numeric YYYYMMDD.}
}
\value{
The converted date variable.
}
\description{
This function converts a RSA-911 date variable written in order YYYYMMDD as
one number to the appropriate date.
}
