% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_sex}
\alias{handle_sex}
\title{Clean sex/gender variable.}
\usage{
handle_sex(x, convert_sex = FALSE)
}
\arguments{
\item{x}{A sex/gender variable, with values 1, 2, 9 (or NA), or 1, 0, 9.}

\item{convert_sex}{True or False, default is False.
Converts 2s to 0s if True.}
}
\value{
The converted date variable.
}
\description{
This function cleans a RSA-911 sex/gender variable, based on user input.
}
