% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_splits}
\alias{handle_splits}
\title{Clean variables in RSA-911 dataset with special characters.}
\usage{
handle_splits(data, var_names)
}
\arguments{
\item{data}{An RSA-911 dataset to be cleaned.}

\item{var_names}{A vector of names for variables that contain special
characters.}
}
\value{
The cleaned RSA-911 dataset, with new variables that contain the
cleaned, separated values without special characters.
}
\description{
This function cleans variables with special characters by splitting values
into separate variables that are separated by special characters.
}
