% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_values}
\alias{handle_values}
\title{Clean incorrect values.}
\usage{
handle_values(x, values, blank_value = NA)
}
\arguments{
\item{x}{An RSA-911 variable with specific allowed values.}

\item{values}{A vector of permitted values for the variable.}

\item{blank_value}{The preferred replacement value for incorrect values.
Default is to replace incorrect values to NA.}
}
\value{
The converted date variable.
}
\description{
This function cleans incorrect values for a RSA-911 variable.
}
