% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_year}
\alias{handle_year}
\title{Clean year variable.}
\usage{
handle_year(x)
}
\arguments{
\item{x}{A year variable, containing a 4-digit year somewhere in string.}
}
\value{
The converted year variable, just the digits.
}
\description{
This function cleans a RSA-911 year variable containing a 4-digit year to
the extracted 4-digit year.
}
