% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load and Combine Data}
\usage{
load_data(directory, files = NULL, download_csv = FALSE)
}
\arguments{
\item{directory}{The directory containing the files}

\item{files}{Optional character vector of file names to extract,
eg. c("PY21Q2", "PY22Q3"). Otherwise, NULL, and all appropriate files
from directory will be extracted.}

\item{download_csv}{Defaults to FALSE. If TRUE, will write a csv file
containing loaded and combined data to working R directory.}
}
\value{
A loaded and combined dataframe. Optionally, a csv data file written
to the working directory.
}
\description{
This function reads in and combines sensitive data from a directory.
}
