% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_scores.R
\name{merge_scores}
\alias{merge_scores}
\title{Merge Scores Data with RSA-911 Data}
\usage{
merge_scores(
  quarterly_data,
  scores_data,
  quarterly_id = "Participant_ID",
  scores_id = "Participant_ID"
)
}
\arguments{
\item{quarterly_data}{The cleaned RSA-911 dataset.}

\item{scores_data}{The cleaned scores dataset.}

\item{quarterly_id}{A string for the variable name corresponding to
participant ID in the cleaned RSA-911 dataset. Defaults to
"Participant_ID".}

\item{scores_id}{A string for the variable name corresponding to
participant ID in the cleaned scores dataset. Defaults to "Participant_ID".}
}
\value{
A data frame with one row per participant.
}
\description{
This function merges cleaned Utah scores data with cleaned Utah RSA-911 Data.
}
