% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rsa_simulated}
\alias{rsa_simulated}
\title{A Simulated Subset of RSA-911 Utah Quarterly Data}
\format{
\subsection{\code{rsa_simulated}}{

A data frame with 1988 rows and 518 columns:
\describe{
\item{variables}{For a list and description of all 518 variables, check the
Case-Service-Report-RSA-911-PD-19-03}
...
}
}
}
\source{
\url{https://rsa.ed.gov/performance/rsa-911-policy-directive}
}
\usage{
rsa_simulated
}
\description{
A small example of quarterly data, simulated to represent real IRB-protected
data.
}
\keyword{datasets}
