% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{scores_simulated}
\alias{scores_simulated}
\title{A Simulated Subset of TRT Scores Data}
\format{
\subsection{\code{scores_simulated}}{

A data frame with 1800 rows and 518 columns:
\describe{
\item{variables}{For an understanding of the 32 variables, check the
publication, Enhancing transition outcomes: A toolkit to facilitate
data-driven pre-employment transition services.}
...
}
}
}
\source{
\url{https://rsa.ed.gov/performance/rsa-911-policy-directive}
}
\usage{
scores_simulated
}
\description{
A small example of TRT scores data, simulated to represent real IRB-protected
data.
}
\keyword{datasets}
