% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{separate_disability}
\alias{separate_disability}
\title{Clean primary and secondary disability variables in RSA-911 dataset.}
\usage{
separate_disability(df)
}
\arguments{
\item{df}{An RSA-911 dataset to be cleaned.}
}
\value{
The cleaned RSA-911 dataset, with new, separate variables for
primary and secondary cause and impairment, without special characters.
}
\description{
This function cleans primary and secondary disability variables with special
characters by splitting impairment and cause values into separate variables
that are separated by special characters.
}
