% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_densities.R
\name{visualize_densities}
\alias{visualize_densities}
\title{Visualize Densities}
\usage{
visualize_densities(
  cat_var,
  num_var,
  cat_var_name = "Categorical Variable",
  num_var_name = "Numeric Variable",
  level_labels = NULL,
  xlab = NULL,
  main = NULL,
  colors = NULL
)
}
\arguments{
\item{cat_var}{The categorical variable.}

\item{num_var}{The numerical variable.}

\item{cat_var_name}{The desired name/label of the categorical variable. If no
name is provided, it will default to "Categorical Variable."}

\item{num_var_name}{The desired name/label of the numeric variable. If no
name is provided, it will default to "Numeric Variable."}

\item{level_labels}{The desired labels of the levels of the categorical
variable. If no labels are provided, the labels will be the values found in
the data.}

\item{xlab}{The desired label of the x-axis, a character string.}

\item{main}{The desired title, a character string.}

\item{colors}{The desired colors for the densities, a character vector.}
}
\value{
A density plot of the inputted variables.
}
\description{
This function provides a user-friendly way to visualize densities of a
numeric variable across levels of a categorical variable.
}
