% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_scores.R
\name{visualize_scores}
\alias{visualize_scores}
\title{Visualize TRT Scores Data}
\usage{
visualize_scores(
  data,
  option = c("overview", "across_service", "across_provider"),
  one_window = FALSE
)
}
\arguments{
\item{data}{A cleaned TRT scores dataset. Apply clean_scores function first.}

\item{option}{The selected visual analysis option. The options are
"overview":
"across_service":
"across_provider":}

\item{one_window}{Whether or not the user wants the visuals to be displayed
in one plotting window, or spread across multiple plot windows. Defaults to
FALSE.}
}
\value{
The appropriate plots for the chosen visual analysis.
}
\description{
This function produces tailored visualizations for a TRT scores dataset in
one step.
}
