% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_provider.R
\name{clean_provider}
\alias{clean_provider}
\title{Clean Scores Data, with Provider Variables Focus}
\usage{
clean_provider(data, condense = FALSE)
}
\arguments{
\item{data}{The scores dataset.}

\item{condense}{Defaults to FALSE. When true, take medians across
participants.}
}
\value{
A cleaned data frame, including only provider-relevant variables,
restructured.
}
\description{
This function cleans a scores dataset, based on the standard data
structure, returning a dataset focused on provider-variable analysis.
}
