% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_scores.R
\name{clean_scores}
\alias{clean_scores}
\title{Clean Scores Data}
\usage{
clean_scores(data, aggregate = TRUE)
}
\arguments{
\item{data}{The scores dataset.}

\item{aggregate}{TRUE or FALSE. Defaults to TRUE, when TRUE, rows are
aggregated to include only unique participants are kept.}
}
\value{
A cleaned data frame, restructured to a wide format, to help with
merging process.
}
\description{
This function cleans a scores dataset, based on the standard data
structure.
}
