% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_utah.R
\name{clean_utah}
\alias{clean_utah}
\title{Clean Utah RSA-911 Data}
\usage{
clean_utah(
  data,
  aggregate = TRUE,
  unidentified_to_0 = TRUE,
  clean_specials = FALSE,
  remove_desc = TRUE,
  remove_strictly_na = TRUE
)
}
\arguments{
\item{data}{The Utah RSA-911 dataset.}

\item{aggregate}{TRUE or FALSE. Defaults to TRUE, when TRUE, rows are
aggregated to include only unique combinations of participants, year,
and quarter.}

\item{clean_specials}{TRUE or FALSE. Defaults to FALSE, when TRUE, variables
containing special characters are cleaned and separated into new variables.}

\item{remove_desc}{TRUE or FALSE. Defaults to TRUE, when TRUE, description
variables are removed to minimize redundancy.}

\item{remove_strictly_na}{TRUE or FALSE. Defaults to TRUE, when TRUE,
variables that contain only NA values are removed.}

\item{unidentified_to_O}{TRUE or FALSE. Defaults to TRUE, when TRUE,
variables where unidentified is represented by 9,
values are converted to 0.}
}
\value{
A cleaned data frame, with incorrect, blank, messy values replaced,
additional, helpful variables created, and unnecessary variables removed.
}
\description{
This function cleans a Utah RSA-911 dataset, based on the standard data
structure.
}
