% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_splits}
\alias{handle_splits}
\title{Clean variables in RSA-911 dataset with special characters.}
\usage{
handle_splits(var, var_name, sep = ";")
}
\arguments{
\item{var}{An RSA-911 special character variable (vector) to be cleaned.}

\item{var_name}{A character vector of the name of the variable to be cleaned.}

\item{sep}{The special character that is separating values within the
variable.}
}
\value{
The new variables derived from the inputted variable. They contain
the cleaned, separated values without special characters.
}
\description{
This function cleans variables with special characters by splitting values
into separate variables that are separated by special characters.
}
